+++ src/main/java/id/myevent/model/dto/EventDto.java
@@ -7,14 +7,14 @@ import lombok.Data;
 public class EventDto {
   private String name;
   private String description;
-  private Integer dateTimeEventStart;
-  private Integer dateTimeEventEnd;
+  private Long dateTimeEventStart;
+  private Long dateTimeEventEnd;
   private String venue;
   private byte[] bannerPhoto;
   private String bannerPhotoName;
   private String bannerPhotoType;
-  private Integer dateTimeRegistrationStart;
-  private Integer dateTimeRegistrationEnd;
+  private Long dateTimeRegistrationStart;
+  private Long dateTimeRegistrationEnd;
   private Long eventStatusId;
   private Long eventCategoryId;
   private Long eventVenueCategoryId;
